backend/requirements.txt

flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
google-generativeai==0.8.3  
groq==0.11.0                
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2